Renamed configuration classes
