+++ test/common/src/main/java/io/camunda/zeebe/spring/test/ZeebeTestConfiguration.java
@@ -1,16 +1,13 @@
 package io.camunda.zeebe.spring.test;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import io.camunda.zeebe.client.api.JsonMapper;
-import io.camunda.zeebe.client.impl.ZeebeObjectMapper;
-import io.camunda.zeebe.spring.client.AbstractZeebeBaseClientSpringConfiguration;
+import io.camunda.zeebe.spring.client.AbstractZeebeBaseConfiguration;
 import io.camunda.zeebe.spring.client.SpringZeebeAutoConfiguration;
 import io.camunda.zeebe.spring.client.testsupport.SpringZeebeTestContext;
+import io.camunda.zeebe.spring.test.configuration.ZeebeTestDefaultConfiguration;
 import io.camunda.zeebe.spring.test.proxy.TestProxyConfiguration;
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.AutoConfigureBefore;
 import org.springframework.boot.autoconfigure.ImportAutoConfiguration;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.jackson.JacksonAutoConfiguration;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
@@ -20,7 +17,7 @@ import org.springframework.context.annotation.Bean;
 @ImportAutoConfiguration({TestProxyConfiguration.class, ZeebeTestDefaultConfiguration.class})
 @AutoConfigureBefore(SpringZeebeAutoConfiguration.class)
 @AutoConfigureAfter(JacksonAutoConfiguration.class) // make sure Spring created ObjectMapper is preferred if available
-public class ZeebeClientTestConfiguration extends AbstractZeebeBaseClientSpringConfiguration {
+public class ZeebeTestConfiguration extends AbstractZeebeBaseConfiguration {
 
   @Bean
   public SpringZeebeTestContext enableTestContext() {
