+++ test/common/src/main/java/io/camunda/zeebe/spring/test/configuration/ZeebeTestDefaultConfiguration.java
@@ -1,4 +1,4 @@
-package io.camunda.zeebe.spring.test;
+package io.camunda.zeebe.spring.test.configuration;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.camunda.zeebe.client.api.JsonMapper;
@@ -6,7 +6,7 @@ import io.camunda.zeebe.client.impl.ZeebeObjectMapper;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Bean;
 
-import static io.camunda.zeebe.spring.client.AbstractZeebeBaseClientSpringConfiguration.DEFAULT_OBJECT_MAPPER;
+import static io.camunda.zeebe.spring.client.AbstractZeebeBaseConfiguration.DEFAULT_OBJECT_MAPPER;
 
 /**
  * Fallback values if certain beans are missing
