+++ test/embedded/src/main/java/io/camunda/zeebe/spring/test/ZeebeSpringTest.java
@@ -1,7 +1,6 @@
 package io.camunda.zeebe.spring.test;
 
 import org.springframework.boot.autoconfigure.ImportAutoConfiguration;
-import org.springframework.context.annotation.Import;
 import org.springframework.test.context.TestExecutionListeners;
 
 import java.lang.annotation.ElementType;
@@ -17,7 +16,7 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.RUNTIME)
 @Inherited
 // this creates the engine and the client
-@ImportAutoConfiguration({ZeebeClientTestConfiguration.class})
+@ImportAutoConfiguration({ZeebeTestConfiguration.class})
 // this listener hooks up into test execution
 @TestExecutionListeners(listeners = ZeebeTestExecutionListener.class, mergeMode = TestExecutionListeners.MergeMode.MERGE_WITH_DEFAULTS)
 public @interface ZeebeSpringTest {
