+++ test/testcontainer/src/main/java/io/camunda/zeebe/spring/test/ZeebeSpringTest.java
@@ -16,7 +16,7 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.RUNTIME)
 @Inherited
 // this creates the engine and the client
-@Import({ZeebeClientTestConfiguration.class})
+@Import({ZeebeTestConfiguration.class})
 // this listener hooks up into test execution
 @TestExecutionListeners(listeners = ZeebeTestExecutionListener.class, mergeMode = TestExecutionListeners.MergeMode.MERGE_WITH_DEFAULTS)
 public @interface ZeebeSpringTest {
